+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -185,6 +185,7 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             dialogBackgroundSchemeColor: controller.dialogBackgroundSchemeColor,
             dialogRadius: controller.dialogBorderRadius,
             timePickerDialogRadius: controller.dialogBorderRadius,
+            dialogElevation: controller.dialogElevation,
             // SnackBar settings.
             snackBarBackgroundSchemeColor: controller.snackBarSchemeColor,
             // AppBar settings.