+++ example/lib/shared/const/store.dart
@@ -890,6 +890,12 @@ class Store {
   // We use NULL as default, on nullable settings.
   static const double? defaultDialogBorderRadius = null;
 
+  // Key used to read and save the dialogElevation value.
+  static const String keyDialogElevation = 'dialogElevation';
+  // Default value for the dialogElevation, also used to reset settings.
+  // We use NULL as default, on nullable settings.
+  static const double? defaultDialogElevation = null;
+
   // Custom surface tint color SETTINGS.
   // ===========================================================================